pyproject.toml
[project]
name = "helm-agent"
version = "0.0.1"
description = "Private-deployment, Hermes-inspired agent backbone with TaskGraph, profiles, skills, and delegation."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "httpx>=0.27.0",
  "openai>=1.0.0",
  "PyYAML>=6.0.1",
  "uvicorn>=0.27.0",
]

[project.optional-dependencies]
web = [
  "fastapi>=0.110.0",
  "uvicorn>=0.27.0",
]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["helm*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--basetemp=.pytest-tmp"
